fix(std.string): to_long 64-bit truncation + to_float garbage
Two silent data-corruption bugs in the string-to-number parse family,
surfaced when a downstream rewrite stashed a 64-bit pointer through
to_long and got it back mangled.

to_long truncated to 32 bits: the wrapper used an inferred multi-return
(`-> {`), which collapses the `long` value slot to 32-bit `int`, so
to_long("4294967296") == 0. Fixed with an explicit `-> (long, string)`
return type (the idiom used elsewhere for non-int multi-returns). The
underlying compiler limitation — inferred multi-return not distinguishing
long from int — is noted in the CHANGELOG for a separate look; the C path
(string_get_long) and the raw extern (string_to_long_raw) were always
correct, which is why string_to_long_raw worked as a workaround.

to_float returned garbage (e.g. 5.3e-315 for "2.5"): string_get_float
returned a 32-bit C `float`, but Aether's `float` is 64-bit (lowers to C
double), so the result register was misread. Now returns (double)v —
parsing still at float precision (to_float's contract; to_double for
full). Found while pinning the family; to_int and to_double were already
correct (to_double's string_get_double already returns C double).

Regression test (tests/regression/test_string_to_number_widths.ae)
covers all four: to_long 64-bit incl. sign, to_int, to_float, to_double,
and invalid-input rejection.

Verified: build; 226 C unit tests; new test + existing string tests pass.

+++ std/string/aether_string.c
@@ -721,8 +721,14 @@ long string_get_long(const void* s) {
 int string_try_float(const void* s) {
     float v; return string_to_float_raw(s, &v);
 }
-float string_get_float(const void* s) {
-    float v = 0.0f; string_to_float_raw(s, &v); return v;
+// Returns double, not float: Aether's `float` type is 64-bit (it lowers
+// to C `double`), so the extern `string_get_float -> float` expects a
+// 64-bit return. Returning a 32-bit C `float` here left Aether reading
+// the result register as a double → garbage (e.g. to_float("2.5") came
+// back 5.3e-315). We still parse at float precision (to_float's contract;
+// use to_double for full precision), then widen for the ABI.
+double string_get_float(const void* s) {
+    float v = 0.0f; string_to_float_raw(s, &v); return (double)v;
 }
 int string_try_double(const void* s) {
     double v; return string_to_double_raw(s, &v);
